Refactored code so there are no errors or warnings.  Changed code to be a static framework lib and an exe game.

+++ Source/Framework/Application.cpp
@@ -113,7 +113,7 @@ void App::Run()
 		double deltaTime = m_glfwTime - previousTime;
 		previousTime = m_glfwTime;
 
-		GameLoop(deltaTime);
+		GameLoop((float)deltaTime);
 
 #ifdef DEBUG
 		ShowFPS(float(deltaTime));
