+++ Source/Game/BrickBreak/BBPaddle.cpp
@@ -20,7 +20,7 @@ void BBPaddle::Init()
 	GameObject::Init();
 }
 
-void BBPaddle::Update(double delta)
+void BBPaddle::Update(float delta)
 {
 	m_Position.x += m_Direction * delta * 100.0f;
