+++ Source/Game/BrickBreak/Brick.h
@@ -7,7 +7,7 @@ public:
 	Brick(class Ball* ball, int idx);
 
 	void Init() override;
-	void Update(double delta) override;
+	void Update(float delta) override;
 
 private:
 	class Ball* m_pBall = nullptr;
